Add files via upload
This dashboard which helps answer the following questions:
What is the most represented gender among loan applicants by country?
How does the loan amount for a select group of countries compare to the global total?
How are loan amounts distributed by sector? Globally? For the select group of countries?
What is the gender split in loan applicants for the select group of countries?